Relocate arrow improvements to the right module
