+++ kotlintest-assertions/kotlintest-assertions-arrow/build.gradle
@@ -6,14 +6,21 @@ plugins {
     id 'signing'
 }
 
+apply plugin: 'kotlin-kapt'
+
 repositories {
     mavenCentral()
 }
 
 dependencies {
     compile project(":kotlintest-assertions")
-    compile 'io.arrow-kt:arrow-typeclasses:0.8.2'
-    compile 'io.arrow-kt:arrow-data:0.8.2'
+    compile "io.arrow-kt:arrow-extras-extensions:$arrow_version"
+    compile "io.arrow-kt:arrow-validation:$arrow_version"
+    compile "io.arrow-kt:arrow-generic:$arrow_version"
+    compile "io.arrow-kt:arrow-effects-io-extensions:$arrow_version"
+    kapt "io.arrow-kt:arrow-meta:$arrow_version"
+    kaptTest "io.arrow-kt:arrow-meta:$arrow_version"
+    testCompile project(":kotlintest-runner:kotlintest-runner-junit5")
 }
 
 apply from: '../../publish.gradle'
